Display messages returned by GStreamer when playing or stopping a stream fails. Wayland support.
[GB.MEDIA.FORM]
* NEW: Wayland support.
* NEW: Display messages returned by GStreamer when playing or stopping a stream fails.

[GB.GTK3]
* BUG: Support for getting the handle of a Wayland window correctly.

[GB.GTK]
* BUG: Support for getting the handle of a Wayland window correctly.

[GB.QT4]
* BUG: Support for getting the handle of a Wayland window correctly.

[GB.QT5]
* BUG: Support for getting the handle of a Wayland window correctly.

+++ gb.gtk/src/CWidget.cpp
@@ -358,7 +358,7 @@ CWIDGET *GetContainer(CWIDGET *control)
 	return GetObject(cont);
 }
 
-int CWIDGET_get_handle(void *_object)
+uintptr_t CWIDGET_get_handle(void *_object)
 {
 	return CONTROL->handle();
 }
